requirements.txt
# 核心依赖
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.9
matplotlib==3.8.0
numpy>=1.25.0
pandas>=2.0.0
alpaca-trade-api==3.1.1

# AI 引擎依赖
google-generativeai==0.3.1  # Google Gemini
openai==1.30.0  # OpenAI GPT

# 数据处理依赖
pytz==2023.3
python-dateutil==2.8.2

# 工具依赖
click==8.1.7
colorama==0.4.6